tmp/requirements.txt
Django==5.1.2
gunicorn==20.0.4
psycopg2
djangorestframework==3.14.0
python-telegram-bot
numpy
matplotlib
pyshorteners
pandas
validators
requests
base58
gspread
gspread-formatting
uvicorn
aiohttp
APScheduler